requirements.txt
pytesseract
poppler-utils
unstructured[all-docs]
pillow 
pydantic 
lxml 
pillow 
matplotlib 
chromadb 
tiktoken 
watermark
python-dotenv
langchain
langchain_community
langchain_google_genai
google-generativeai

openai 
